solution

+++ 1327A.cpp
@@ -4,7 +4,7 @@ using namespace std;
 int main(){
 	int tc;cin>>tc;
 	while(tc--){
-		int n,k; cin>>n>>k;
+		long long n,k; cin>>n>>k;
 		if (n%2==k%2 && k*k<=n)
 		{
 			cout<<"YES"<<endl;
